requirements.txt
langchain==0.3.1
langchain-community==0.3.1
langchain-groq==0.2.0
langchain-openai==0.2.1
langchain-chroma==0.1.4
Faker==30.0.0
python-dotenv==1.0.1
streamlit==1.38.0
pydpf==5.1.0
sentence-transformers==3.3.1
langchain_huggingface==0.1.2
faiss-cpu==1.9.0.post1